runpod-worker/requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
ase
pymatgen
huggingface_hub
fairchem-core
orb_models